+++ demos/editor_pen_multiple/Form1.cs
@@ -122,9 +122,6 @@ namespace Demos
             pen.PenColor = color; //Config.EntityPenColors[0]; //default 'white'
             pen.Description = color.ToString();
 
-            pen.LaserFpk = 0;
-            pen.LaserQSwitchDelay = 0;
-
             pen.Power = 1;
             pen.Frequency = 50 * 1000;
             pen.PulseWidth = 2;
@@ -143,6 +140,8 @@ namespace Demos
             pen.RasterDirection = EntityPen.RasterDirections.Horizontal;
             pen.IsRasterZigZag = true;
             pen.PixelTime = 100;
+            pen.PixelPulses = 0;
+            pen.IsPixelPulsesExit = true;
             pen.PixelPeriod = 200;
             pen.PixelChannel = ExtensionChannels.ExtAO2;