+++ demos/editor_ui/SiriusMultiEditorControl.cs
@@ -780,6 +780,7 @@ namespace Demos
                 splitContainer2.Panel2Collapsed = !splitContainer2.Panel2Collapsed;
             };
 
+            Document = new DocumentBase();
         }
 
         /// <summary>
@@ -927,11 +928,7 @@ namespace Demos
         /// <param name="e">An <see cref="EventArgs"/> that contains the event data.</param>
         private void SiriusEditorControl_Load(object sender, EventArgs e)
         {
-            Document = new DocumentBase();
-            //Document.ActNew(true, true, true, true, true, true, true, true, true);
-            Document.ActNew(true, true, true, true, true, false, false, false, false); //scanner layer 펜 재 생성 막기
-
-            SwitchDevices(0);
+            SwitchDevices(CurrentDeviceIndex);
         }
         /// <summary>
         /// Handles form closing; disposes timers.